Add solving-sudoku as a submodule
The first goal is to keep the code in rating-sudoku repository
in line with the output of solving-sudoku transformations,
and along their planned evolutions.

As an added bonus, it will allow me to extend the script rating.sh
to load a sudoku file, convert it to a numeric value and trigger
the solver to produce the result file before conversion for rating.